requirements.txt
streamlit>=1.29.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
huggingface-hub>=0.19.4
langchain>=0.0.350
beautifulsoup4>=4.12.2
requests>=2.31.0